+++ DoomerBot.py
@@ -2,8 +2,7 @@ import datetime
 from discord.ext import tasks
 import discord
 from discord import app_commands
-import function_controller
-import responses
+import FunctionController
 
 
 class DoomerBot(discord.Client):
@@ -17,22 +16,22 @@ class DoomerBot(discord.Client):
 
     async def on_ready(self):
         """
-        # pro nasazování commandů jen na specificke servery, nefunguje, možná se bude hodit idk
+        # For inserting commands to specific servers, currently not working, might use later.
         ids = []
         for guild in self.guilds:
             ids.append(int(guild.id))
         self.tree.copy_global_to(guilds=discord.Object(id=ids)) # for multiple guilds need to change guild to guilds
         """
-        # self.tree.clear_commands(guild=None) # pro vymazani commandu
+        # self.tree.clear_commands(guild=None) # for purging commands from servers
         # await self.tree.sync(guild=None)
         await self.tree.sync()
         print(f'Logged in as {self.user} (ID: {self.user.id})')
 
     @tasks.loop(time=datetime.time(hour=19, minute=00))  # task runs every day at 19:00 utc = 20:00 czech time
     async def daily_feed_send(self):
-        for channel_id in function_controller.load_channel_id():
+        for channel_id in FunctionController.load_channel_id():
             channel = self.get_channel(channel_id)
-            picture = responses.handle_response("hentai", "doomer")
+            picture = FunctionController.handle_response("hentai", "doomer")
             await channel.send(f"Your daily waifu!\n{picture}")
 
     @daily_feed_send.before_loop
